+++ templates/admin-panel/template/client/__tests__/integration/helpers.js
@@ -2,6 +2,7 @@ import { start, DI } from 'sham-ui';
 import pretty from 'pretty';
 import initializer from '../../src/initializers/main';
 import RoutesMembersPage from '../../src/components/routes/members/page.sfc';
+import RoutesServerInfoPage from '../../src/components/routes/server-info/page.sfc';
 
 export const app = {
     async start( waitRendering = true ) {
@@ -84,6 +85,6 @@ export default function() {
     clearBody();
     resetStorage();
     setupRouter();
-    setupLazyPageMocks( RoutesMembersPage );
+    setupLazyPageMocks( RoutesMembersPage, RoutesServerInfoPage );
     Object.defineProperty( window, 'CSS', { value: () => ( {} ) } );
 }