+++ templates/admin-panel/template/client/__tests__/integration/redirects/non-superuser-cant-visit-server.js
@@ -1,0 +1,20 @@
+import setup, { app } from '../helpers';
+import axios from 'axios';
+jest.mock( 'axios' );
+
+beforeEach( () => {
+    jest.resetModules();
+    jest.clearAllMocks();
+    setup();
+} );
+
+it( 'server', async() => {
+    expect.assertions( 2 );
+
+    axios.useDefaultMocks();
+
+    history.pushState( {}, '', 'http://client.example.com/server/' );
+    await app.start();
+    app.checkBody();
+    expect( window.location.href ).toBe( 'http://client.example.com/' );
+} );